Move test_vector_builder into separate file
